ui_requirements.txt
absl-py==0.9.0
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.7
astor==0.8.1
attrs==19.3.0
auto-py-to-exe==2.7.5
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.3.0
bottle==0.12.18
bottle-websocket==0.2.9
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.7
Eel==0.12.4
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.2.2
gevent==20.6.2
gevent-websocket==0.10.1
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.29.0
h5py==2.10.0
idna==2.9
image==1.5.32
imageio==2.8.0
imbalanced-learn==0.7.0
imblearn==0.0
imgaug==0.4.0
importlib-metadata==1.6.1
imutils==0.5.3
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.2.0
lxml==4.6.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
PyInstaller==3.6
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.15.0.1.7
pyrsistent==0.16.0
PySide2==5.15.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.6.8
removebg==0.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.0
shiboken2==5.15.0
six==1.15.0
sockets==1.0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
systemtools==1.0.2
tensorboard==2.1.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.3.1
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
tornado==6.0.3
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.18
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.0
zipp==2.2.0
zope.event==4.4
zope.interface==5.1.0
